Create ascii_orig.mif
It is the mif format of the characters. It stores all the characters. A character occupies 8 horizontal and 16 vertical pixels on the screen. Here each location has stored the one set of 8 pixels. Each pixel has taken 1 bit of information- either on or off. It glows on being otherwise off.
location: information
You can observe a pattern of the characters like '2' from location number 799 to 814 represented using 1's.